Phase 4 Wave A: move CreatorStudioView.tsx into surfaces/studio/
Final batch of Creator Studio's core/+surfaces/ scaffolding (code editor
was #1083, widget panels #1084). CreatorStudioView.tsx was the natural
last move — everything it renders now already lives in surfaces/studio/,
so this batch is just the one file plus its own six test files, all
imports repointed.

studioApi.ts, studioShelf.ts, submissionApi.ts, editorBridge.ts, and the
other root-level primitives it uses stay in place, same as every prior
Studio batch. App.tsx's lazy route import is repointed at the new path.

The whole Studio surface is now consolidated under
apps/web/src/surfaces/studio/.
