+++ apps/web/src/surfaces/studio/CreatorStudioView.theater.test.tsx
@@ -4,9 +4,9 @@ import { act, createElement } from 'react';
 import { createRoot } from 'react-dom/client';
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 import { CreatorStudioView } from './CreatorStudioView.js';
-import i18n from './i18n/index.js';
-import type { StudioGamesResponse } from './studioApi.js';
-import { getSubmissionPreview, getSubmissionStatus } from './submissionApi.js';
+import i18n from '../../i18n/index.js';
+import type { StudioGamesResponse } from '../../studioApi.js';
+import { getSubmissionPreview, getSubmissionStatus } from '../../submissionApi.js';
 
 /**
  * The site's full `GameTheater` — fullscreen, share, report — reachable from Studio's
@@ -22,12 +22,12 @@ const fetchStudioScorecards = vi.fn();
 const fetchStudioSuggestions = vi.fn();
 let authUser: { uid: string; name: string } | null = null;
 
-vi.mock('./AuthContext', () => ({
+vi.mock('../../AuthContext', () => ({
   useAuth: () => ({ user: authUser, logout: vi.fn() }),
 }));
 
-vi.mock('./studioApi', async () => {
-  const actual = await vi.importActual<typeof import('./studioApi.js')>('./studioApi.js');
+vi.mock('../../studioApi', async () => {
+  const actual = await vi.importActual<typeof import('../../studioApi.js')>('../../studioApi.js');
   return {
     ...actual,
     fetchStudioGames: (...args: unknown[]) => fetchStudioGames(...args),
@@ -38,8 +38,8 @@ vi.mock('./studioApi', async () => {
   };
 });
 
-vi.mock('./submissionApi', async () => {
-  const actual = await vi.importActual<typeof import('./submissionApi.js')>('./submissionApi.js');
+vi.mock('../../submissionApi', async () => {
+  const actual = await vi.importActual<typeof import('../../submissionApi.js')>('../../submissionApi.js');
   return {
     ...actual,
     getSubmissionStatus: vi.fn(),
