+++ apps/web/src/surfaces/studio/CreatorStudioView.tsx
@@ -1,29 +1,29 @@
 import type { DismissReason } from '@gamedevpl/contract';
 import { useEffect, useId, useMemo, useRef, useState, type RefObject } from 'react';
 import { useTranslation } from 'react-i18next';
-import { useAuth } from './AuthContext.js';
-import { AuthModal } from './AuthModal.js';
-import { ClaimHandleModal } from './ClaimHandleModal.js';
-import { StudioCreatorProfileProvider } from './studioCreatorProfile.js';
-import type { GameHealth } from './healthApi.js';
-import { PixelIcon, type PixelIconName } from './PixelIcon.js';
-import { formatRelativeTime } from './relativeTime.js';
-import { playPath, studioPath, type StudioTab } from './router.js';
-import { abandonSubmission, deleteGame, handoffToPlatform } from './submissionApi.js';
-import { StudioShotToasts } from './surfaces/studio/StudioShotToasts.js';
-import { type CodeActionsMode } from './surfaces/studio/CodeActionsMenu.js';
-import { CodeSurface } from './surfaces/studio/CodeSurface.js';
-import { EditorPanel } from './surfaces/studio/EditorPanel.js';
-import { StudioStage, type StagePosture, type StageStatus } from './surfaces/studio/StudioStage.js';
-import { StudioStrip } from './surfaces/studio/StudioStrip.js';
-import { usePlayChromeIdle } from './usePlayChromeIdle.js';
-import { StudioVersionRibbon } from './surfaces/studio/StudioVersionRibbon.js';
-import { StudioChatRail } from './surfaces/studio/StudioChatRail.js';
-import { StudioStageCard } from './surfaces/studio/StudioStageCard.js';
-import { StudioFullBleed } from './surfaces/studio/StudioFullBleed.js';
-import { useStageSource, type StageOrigin } from './useStageSource.js';
-import { useStudioStatusPoll, defaultRailOpen } from './surfaces/studio/useStudioStatusPoll.js';
-import { GameTheater } from './GameTheater.js';
+import { useAuth } from '../../AuthContext.js';
+import { AuthModal } from '../../AuthModal.js';
+import { ClaimHandleModal } from '../../ClaimHandleModal.js';
+import { StudioCreatorProfileProvider } from '../../studioCreatorProfile.js';
+import type { GameHealth } from '../../healthApi.js';
+import { PixelIcon, type PixelIconName } from '../../PixelIcon.js';
+import { formatRelativeTime } from '../../relativeTime.js';
+import { playPath, studioPath, type StudioTab } from '../../router.js';
+import { abandonSubmission, deleteGame, handoffToPlatform } from '../../submissionApi.js';
+import { StudioShotToasts } from './StudioShotToasts.js';
+import { type CodeActionsMode } from './CodeActionsMenu.js';
+import { CodeSurface } from './CodeSurface.js';
+import { EditorPanel } from './EditorPanel.js';
+import { StudioStage, type StagePosture, type StageStatus } from './StudioStage.js';
+import { StudioStrip } from './StudioStrip.js';
+import { usePlayChromeIdle } from '../../usePlayChromeIdle.js';
+import { StudioVersionRibbon } from './StudioVersionRibbon.js';
+import { StudioChatRail } from './StudioChatRail.js';
+import { StudioStageCard } from './StudioStageCard.js';
+import { StudioFullBleed } from './StudioFullBleed.js';
+import { useStageSource, type StageOrigin } from '../../useStageSource.js';
+import { useStudioStatusPoll, defaultRailOpen } from './useStudioStatusPoll.js';
+import { GameTheater } from '../../GameTheater.js';
 import {
   collapseStudioGames,
   filterStudioGames,
@@ -35,16 +35,16 @@ import {
   STUDIO_SHELF_TOOLS_AT,
   type StudioShelfFilter,
   type StudioShelfGame,
-} from './studioShelf.js';
-import { StudioConnectCard } from './surfaces/studio/StudioConnectCard.js';
-import { StudioCreatorAgentKeyPanel } from './surfaces/studio/StudioCreatorAgentKeyPanel.js';
-import { StudioDetailsBuildProgress } from './surfaces/studio/StudioDetailsBuildProgress.js';
-import { StudioDetailsMedia } from './surfaces/studio/StudioDetailsMedia.js';
-import { ContributionsSetting } from './ContributionsSetting.js';
-import { ProposalReviewPanel } from './surfaces/review/ProposalReviewPanel.js';
-import { StudioOAuthClientsPanel } from './surfaces/studio/StudioOAuthClientsPanel.js';
-import { StudioWorkspaceCheckoutPanel } from './surfaces/studio/StudioWorkspaceCheckoutPanel.js';
-import { SubmissionStatusView } from './surfaces/studio/SubmissionStatusView.js';
+} from '../../studioShelf.js';
+import { StudioConnectCard } from './StudioConnectCard.js';
+import { StudioCreatorAgentKeyPanel } from './StudioCreatorAgentKeyPanel.js';
+import { StudioDetailsBuildProgress } from './StudioDetailsBuildProgress.js';
+import { StudioDetailsMedia } from './StudioDetailsMedia.js';
+import { ContributionsSetting } from '../../ContributionsSetting.js';
+import { ProposalReviewPanel } from '../review/ProposalReviewPanel.js';
+import { StudioOAuthClientsPanel } from './StudioOAuthClientsPanel.js';
+import { StudioWorkspaceCheckoutPanel } from './StudioWorkspaceCheckoutPanel.js';
+import { SubmissionStatusView } from './SubmissionStatusView.js';
 import {
   approveSuggestion,
   dismissSuggestion,
@@ -59,8 +59,8 @@ import {
   type StudioScorecard,
   type StudioSuggestion,
   type AutonomyMode,
-} from './studioApi.js';
-import type { EditorContentPush, EditorControllerState } from './editorBridge.js';
+} from '../../studioApi.js';
+import type { EditorContentPush, EditorControllerState } from '../../editorBridge.js';
 
 /**
  * Creator control panel (docs/improvement-loop-plan.md IL-2 creator surface).
