+++ eslint-rules/comment-prose-baseline.json
@@ -397,10 +397,6 @@
     "apps/web/src/CreatorProfilePage.tsx": 105,
     "apps/web/src/CreatorQA.test.tsx": 526,
     "apps/web/src/CreatorQA.tsx": 1041,
-    "apps/web/src/CreatorStudioView.handoff.test.tsx": 291,
-    "apps/web/src/CreatorStudioView.test.tsx": 858,
-    "apps/web/src/CreatorStudioView.theater.test.tsx": 69,
-    "apps/web/src/CreatorStudioView.tsx": 2931,
     "apps/web/src/editorBridge.ts": 228,
     "apps/web/src/EditorCollections.test.tsx": 13,
     "apps/web/src/editorContentTools.ts": 147,
@@ -563,6 +559,10 @@
     "apps/web/src/surfaces/studio/codeSurfaceApi.ts": 82,
     "apps/web/src/surfaces/studio/codeTokens.ts": 177,
     "apps/web/src/surfaces/studio/connectApi.ts": 231,
+    "apps/web/src/surfaces/studio/CreatorStudioView.handoff.test.tsx": 291,
+    "apps/web/src/surfaces/studio/CreatorStudioView.test.tsx": 858,
+    "apps/web/src/surfaces/studio/CreatorStudioView.theater.test.tsx": 69,
+    "apps/web/src/surfaces/studio/CreatorStudioView.tsx": 2931,
     "apps/web/src/surfaces/studio/EditorCollections.test.tsx": 13,
     "apps/web/src/surfaces/studio/editorJsonLiveDiff.ts": 92,
     "apps/web/src/surfaces/studio/EditorPanel.test.tsx": 16,
