+++ eslint-rules/module-size-baseline.json
@@ -484,7 +484,7 @@
     "apps/web/src/App.gamePageTheater.test.ts": 250,
     "apps/web/src/App.overlayExit.test.ts": 161,
     "apps/web/src/App.qa.test.ts": 484,
-    "apps/web/src/App.tsx": 1358,
+    "apps/web/src/App.tsx": 1360,
     "apps/web/src/AppleSignInButton.test.tsx": 120,
     "apps/web/src/AppleSignInButton.tsx": 149,
     "apps/web/src/AppLoadingScreen.tsx": 18,
@@ -543,12 +543,6 @@
     "apps/web/src/CreatorProfilePage.tsx": 206,
     "apps/web/src/CreatorQA.test.tsx": 577,
     "apps/web/src/CreatorQA.tsx": 598,
-    "apps/web/src/CreatorStudioView.claimGate.test.tsx": 140,
-    "apps/web/src/CreatorStudioView.codeActions.test.tsx": 276,
-    "apps/web/src/CreatorStudioView.handoff.test.tsx": 274,
-    "apps/web/src/CreatorStudioView.test.tsx": 2143,
-    "apps/web/src/CreatorStudioView.theater.test.tsx": 145,
-    "apps/web/src/CreatorStudioView.tsx": 2241,
     "apps/web/src/diffLines.test.ts": 50,
     "apps/web/src/editorBridge.test.ts": 23,
     "apps/web/src/editorBridge.ts": 521,
@@ -816,6 +810,13 @@
     "apps/web/src/surfaces/studio/codeSurfaceUpload.ts": 158,
     "apps/web/src/surfaces/studio/codeTokens.ts": 217,
     "apps/web/src/surfaces/studio/connectApi.ts": 175,
+    "apps/web/src/surfaces/studio/CreatorStudioView.claimGate.test.tsx": 140,
+    "apps/web/src/surfaces/studio/CreatorStudioView.codeActions.test.tsx": 272,
+    "apps/web/src/surfaces/studio/CreatorStudioView.editorSurface.test.ts": 12,
+    "apps/web/src/surfaces/studio/CreatorStudioView.handoff.test.tsx": 274,
+    "apps/web/src/surfaces/studio/CreatorStudioView.test.tsx": 2143,
+    "apps/web/src/surfaces/studio/CreatorStudioView.theater.test.tsx": 145,
+    "apps/web/src/surfaces/studio/CreatorStudioView.tsx": 2238,
     "apps/web/src/surfaces/studio/diffLines.test.ts": 50,
     "apps/web/src/surfaces/studio/diffLines.ts": 46,
     "apps/web/src/surfaces/studio/EditorCollections.test.tsx": 595,
